uek-rpm: Make sure dtb directory exists for emb3.
There is no device tree for Bluefield 3 and no dtb directory.
However rpmbuild or Jenkins will not build the kernel without
it. Previously emb3 builds included support for a number of
platforms, Some if these caused the dtb directory to get added.
The emb3 build has been stripped down, and the dtb directory is
no longer getting added. Add it explicitly for emb3 builds.

Orabug: 37910874

+++ uek-rpm/ol9/kernel-uek.spec
@@ -1228,9 +1228,13 @@ BuildKernel() {
     mkdir -p $RPM_BUILD_ROOT/%{image_install_path}
     mkdir -p $RPM_BUILD_ROOT/lib/modules/$KernelVer
 %ifarch %{arm} aarch64
-    make %{?make_opts} ARCH=$Arch %{?_kernel_cc} dtbs dtbs_install INSTALL_DTBS_PATH=$RPM_BUILD_ROOT/%{image_install_path}/dtb-$KernelVer
-    cp -r $RPM_BUILD_ROOT/%{image_install_path}/dtb-$KernelVer $RPM_BUILD_ROOT/lib/modules/$KernelVer/dtb
-    find arch/$Arch/boot/dts -name '*.dtb' -type f | xargs rm -f
+    if [ "$Flavour" != "emb3" ]; then
+	make %{?make_opts} ARCH=$Arch %{?_kernel_cc} dtbs dtbs_install INSTALL_DTBS_PATH=$RPM_BUILD_ROOT/%{image_install_path}/dtb-$KernelVer
+	cp -r $RPM_BUILD_ROOT/%{image_install_path}/dtb-$KernelVer $RPM_BUILD_ROOT/lib/modules/$KernelVer/dtb
+	find arch/$Arch/boot/dts -name '*.dtb' -type f | xargs rm -f
+    else
+	mkdir $RPM_BUILD_ROOT/lib/modules/$KernelVer/dtb
+    fi
 %endif
 
 %if %{with_ctf}
